Revert "Merge pull request #199 from openSUSE/feature/opensuse"
This reverts commit 97347ec44d927c854b93d6722964aa459e42a106, reversing
changes made to 051564facd852ef5bbb62db9625631a21f2199d5.
