+++ cmd/clair/main.go
@@ -30,9 +30,7 @@ import (
 
 	_ "github.com/coreos/clair/updater/fetchers/alpine"
 	_ "github.com/coreos/clair/updater/fetchers/debian"
-	_ "github.com/coreos/clair/updater/fetchers/opensuse"
 	_ "github.com/coreos/clair/updater/fetchers/rhel"
-	_ "github.com/coreos/clair/updater/fetchers/sle"
 	_ "github.com/coreos/clair/updater/fetchers/ubuntu"
 	_ "github.com/coreos/clair/updater/metadata_fetchers/nvd"
